Move query rewrite to ee and build test framework.
