+++ src/query/ee-features/aggregating-index/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "aggregating-index"
+description = "aggregating index handler"
+version = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+publish = { workspace = true }
+edition = { workspace = true }
+
+[lib]
+doctest = false
+test = false
+
+[dependencies]
+# Workspace dependencies
+common-base = { path = "../../../common/base" }
+common-exception = { path = "../../../common/exception" }
+common-sql = { path = "../../sql" }
+
+[build-dependencies]
+
+[package.metadata.cargo-machete]
+ignored = ["match-template"]
