+++ src/query/ee-features/aggregating-index/src/aggregating_index_handler.rs
@@ -1,0 +1,41 @@
+// Copyright 2021 Datafuse Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::sync::Arc;
+
+use common_base::base::GlobalInstance;
+use common_exception::Result;
+use common_sql::optimizer::SExpr;
+
+pub trait AggregatingIndexHandler: Sync + Send {
+    fn do_query_rewrite(&self, s_expr: &SExpr, indexes: &[SExpr]) -> Result<Option<SExpr>>;
+}
+
+pub struct AggregatingIndexHandlerWrapper {
+    handler: Box<dyn AggregatingIndexHandler>,
+}
+
+impl AggregatingIndexHandlerWrapper {
+    pub fn new(handler: Box<dyn AggregatingIndexHandler>) -> Self {
+        Self { handler }
+    }
+
+    pub fn do_query_rewrite(&self, s_expr: &SExpr, indexes: &[SExpr]) -> Result<Option<SExpr>> {
+        self.handler.do_query_rewrite(s_expr, indexes)
+    }
+}
+
+pub fn get_agg_index_handler() -> Arc<AggregatingIndexHandlerWrapper> {
+    GlobalInstance::get()
+}
