+++ src/query/ee/Cargo.toml
@@ -13,13 +13,16 @@ test = false
 
 [dependencies]
 # Workspace dependencies
+aggregating-index = { path = "../ee-features/aggregating-index" }
 async-backtrace = { workspace = true }
 common-base = { path = "../../common/base" }
 common-catalog = { path = "../catalog" }
 common-config = { path = "../config" }
 common-exception = { path = "../../common/exception" }
+common-expression = { path = "../expression" }
 common-license = { path = "../../common/license" }
 common-settings = { path = "../../query/settings" }
+common-sql = { path = "../sql" }
 common-storages-fuse = { path = "../storages/fuse" }
 databend-query = { path = "../service" }
 storages-common-cache = { path = "../storages/common/cache" }
@@ -31,6 +34,10 @@ chrono = { workspace = true }
 dashmap = "5.4"
 jwt-simple = "0.11.0"
 tracing = "0.1.36"
+
+[dev-dependencies]
+common-ast = { path = "../../query/ast" }
+parking_lot = "0.12.1"
 
 [build-dependencies]
 common-building = { path = "../../common/building" }
