+++ src/query/ee/src/aggregating_index/aggregating_index_handler.rs
@@ -1,0 +1,40 @@
+// Copyright 2023 Databend Cloud
+//
+// Licensed under the Elastic License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.elastic.co/licensing/elastic-license
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::sync::Arc;
+
+use aggregating_index::aggregating_index_handler::AggregatingIndexHandler;
+use aggregating_index::aggregating_index_handler::AggregatingIndexHandlerWrapper;
+use common_base::base::GlobalInstance;
+use common_exception::Result;
+use common_sql::optimizer::SExpr;
+
+use super::query_rewrite;
+
+pub struct RealAggregatingIndexHandler {}
+
+impl AggregatingIndexHandler for RealAggregatingIndexHandler {
+    fn do_query_rewrite(&self, s_expr: &SExpr, indexes: &[SExpr]) -> Result<Option<SExpr>> {
+        query_rewrite::rewrite(s_expr, indexes)
+    }
+}
+
+impl RealAggregatingIndexHandler {
+    pub fn init() -> Result<()> {
+        let rm = RealAggregatingIndexHandler {};
+        let wrapper = AggregatingIndexHandlerWrapper::new(Box::new(rm));
+        GlobalInstance::set(Arc::new(wrapper));
+        Ok(())
+    }
+}
