+++ src/query/ee/src/lib.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub mod aggregating_index;
 pub mod enterprise_services;
 pub mod license;
 pub mod storages;
