+++ src/query/ee/tests/it/aggregating_index/mod.rs
@@ -12,21 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use common_config::InnerConfig;
-use common_exception::Result;
-use common_license::license_manager::LicenseManager;
-
-use crate::aggregating_index::RealAggregatingIndexHandler;
-use crate::license::license_mgr::RealLicenseManager;
-use crate::storages::fuse::operations::RealVacuumHandler;
-
-pub struct EnterpriseServices;
-impl EnterpriseServices {
-    #[async_backtrace::framed]
-    pub async fn init(_config: InnerConfig) -> Result<()> {
-        RealLicenseManager::init()?;
-        RealVacuumHandler::init()?;
-        RealAggregatingIndexHandler::init()?;
-        Ok(())
-    }
-}
+mod query_rewrite;
